+++ starter/app/src/main/java/com/example/android/politicalpreparedness/data/network/models/GetVotersRequest.kt
@@ -2,5 +2,5 @@ package com.example.android.politicalpreparedness.data.network.models
 
 data class GetVotersRequest (
     val address: String,
-    val electionId: Long?
+    val electionId: Int?
 )
